Workinng backend

+++ src/main/java/com/jai/croop/model/BaseDiscount.java
@@ -5,14 +5,14 @@ import jakarta.persistence.MappedSuperclass;
 @MappedSuperclass
 public class BaseDiscount {
     private int originalPrice;
-    private long discountPercent;
+    private double discountPercent;
     private int salePrice;
 
     public BaseDiscount(){
 
     }
 
-    public BaseDiscount(int originalPrice, int discountPercent, int salePrice){
+    public BaseDiscount(int originalPrice, double discountPercent, int salePrice){
         this.originalPrice = originalPrice;
         this.discountPercent = discountPercent;
         this.salePrice = salePrice;
@@ -25,10 +25,10 @@ public class BaseDiscount {
         this.originalPrice = originalPrice;
     }
 
-    public long getDiscountPercent() {
+    public double getDiscountPercent() {
         return discountPercent;
     }
-    public void setDiscountPercent(long discountPercent) {
+    public void setDiscountPercent(double discountPercent) {
         this.discountPercent = discountPercent;
     }
